+++ src/sub/models/polynet/mlj_neural_polynet.jl
@@ -60,9 +60,11 @@ function MLJFlux.build(nn::PolyNet, rng, n_in, n_out)
         )
     end
 
-    return Flux.Chain(
-        Flux.Dense(n_in, nn.neuron, bias = false, init = init),
-        Flux.Chain(inner_layer...),
-        Flux.Dense(nn.neuron, n_out, bias = false, init = init),
+    y = Flux.Chain(
+        polynet_input = Flux.Dense(n_in, nn.neuron, bias = false, init = init),
+        polynet_inner = Flux.Chain(inner_layer...),
+        polynet_output = Flux.Dense(nn.neuron, n_out, bias = false, init = init),
     )
+
+    return y
 end